integration

+++ src/main/java/cc/officina/gatorade/web/rest/SessionResource.java
@@ -1,7 +1,10 @@
 package cc.officina.gatorade.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
+
+import cc.officina.gatorade.domain.Game;
 import cc.officina.gatorade.domain.Session;
+import cc.officina.gatorade.service.GameService;
 import cc.officina.gatorade.service.SessionService;
 import cc.officina.gatorade.web.rest.util.HeaderUtil;
 import cc.officina.gatorade.web.rest.util.PaginationUtil;
@@ -35,9 +38,11 @@ public class SessionResource {
     private static final String ENTITY_NAME = "session";
 
     private final SessionService sessionService;
+    private final GameService gameService;
 
-    public SessionResource(SessionService sessionService) {
+    public SessionResource(SessionService sessionService, GameService gameService) {
         this.sessionService = sessionService;
+        this.gameService = gameService;
     }
 
     /**
@@ -50,9 +55,19 @@ public class SessionResource {
     @PostMapping("/sessions")
     @Timed
     public ResponseEntity<Session> createSession(@RequestBody Session session) throws URISyntaxException {
-        log.debug("REST request to save Session : {}", session);
+        log.info("REST request to save Session : {}", session);
         if (session.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "idexists", "A new session cannot already have an ID")).body(null);
+        }
+        //VALIDATION!
+        Game game = gameService.findOne(session.getGame().getId());
+        if(game == null) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "invalidGameId", "Invalid game id")).body(null);
+        }
+        Session oldSession = sessionService.findOneByExtId(session.getExtId());
+        if(oldSession != null)
+        {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "invalidExtId", "extId already used")).body(null);
         }
         Session result = sessionService.saveAndSchedule(session.getGame(), session);
         return ResponseEntity.created(new URI("/api/sessions/" + result.getId()))
@@ -75,6 +90,21 @@ public class SessionResource {
         log.debug("REST request to update Session : {}", session);
         if (session.getId() == null) {
             return createSession(session);
+        }
+        //VALIDATION
+        Session _session = sessionService.findOne(session.getId());
+        if(session.getMatches() != null && session.getMatches().size() > 0)
+        {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "sessionAlreadyInUse", "Session already in use")).body(null);
+        }
+        Game game = gameService.findOne(session.getGame().getId());
+        if(game == null) {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "invalidGameId", "Invalid game id")).body(null);
+        }
+        Session oldSession = sessionService.findOneByExtId(session.getExtId());
+        if(oldSession != null && !oldSession.getId().equals(session.getId()))
+        {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "invalidExtId", "extId already used")).body(null);
         }
         Session result = sessionService.save(session);
         return ResponseEntity.ok()
@@ -121,6 +151,12 @@ public class SessionResource {
     @Timed
     public ResponseEntity<Void> deleteSession(@PathVariable Long id) {
         log.debug("REST request to delete Session : {}", id);
+        //VALIDATION
+        Session session = sessionService.findOne(id);
+        if(session.getMatches() != null && session.getMatches().size() > 0)
+        {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "sessionAlreadyInUse", "Session already in use")).body(null);
+        }
         sessionService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
@@ -131,8 +167,12 @@ public class SessionResource {
     public ResponseEntity<Session> elaborateSession(@PathVariable Long id) throws URISyntaxException {
         log.debug("REST request to elaborate Session with id " + id);
         Session session = sessionService.findOne(id);
-        if (session.getId() == null) {
+        if (session == null || session.getId() == null) {
         	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("notFound", "SessionNotFound", "Session not found")).body(null);
+        }
+        if(session.isElaborated())
+        {
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("alreadyElaborated", "SessionAlreadyElab", "Session already processed")).body(null);
         }
         sessionService.elaborate(session);
         return ResponseEntity.ok()
